pyproject.toml
[project]
name = "anspg"
version = "0.1.0"
description = "Adversarial Neuro-Symbolic Proving Ground - LLM 1v1 Benchmark for Lean 4"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "ANSPG Team" }
]
keywords = ["lean4", "theorem-proving", "adversarial", "repl", "llm", "benchmark"]

dependencies = [
    # LeanDojo - Optional for static repo analysis (REPL is primary for dynamic verification)
    # "lean-dojo>=1.7.0",
    # LLM clients
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    # Data/ML utilities
    "numpy>=1.26.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.3.0",
    # CLI
    "rich>=13.7.0",
    "typer>=0.9.0",
    # Utilities
    "pydantic>=2.5.0",
    "aiofiles>=23.2.1",
    "networkx>=3.2.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.13.2",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "black>=24.1.0",
]

[project.scripts]
anspg = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
